Refactor _zstd_load_c_dict and _zstd_load_d_dict.
